pyproject.toml
[tool.poetry]
name = "rtspmotion"
version = "0.1.0"
description = ""
authors = ["Joost Nieuwenhuijse <joost@newhouse.nl>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
numpy = "^2.1.2"
paho-mqtt = "^2.1.0"
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
